aparte controller gemaakt voor newMeal
